requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
gunicorn==20.0.4
clear-cache==1.0.0
django-clear-cache==0.3
django-crispy-forms==1.14.0
django-filter==21.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-notifications-latest==2.1.0
django-tables2==2.4.1
jsonfield==3.1.0
mysqlclient==2.1.0
notifications==0.3.2
pytz==2022.1
swapper==1.3.0